\name{HestonEMcpp-package}
\alias{HestonEMcpp-package}
\alias{HestonEMcpp}
\docType{package}
\title{
\packageTitle{HestonEMcpp}
}
\description{
\packageDescription{HestonEMcpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HestonEMcpp}
\packageIndices{HestonEMcpp}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{HestonEMcpp}

Maintainer: \packageMaintainer{HestonEMcpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
